Adds database access and authentication through firebase
I added classes interfaces and retrofit services to connect to and use an online source as a database for reports if we should choose to do user reports that way.

Otherwise I started creating firebase connections and a delay class with options connecting it to firebase to allow for a delay database for user reports and to allow for user authentication.  I just need some testing and to add it to the page for the application.